sampler: Align sampler output file better
This makes everything in the sampler output file 8 byte aligned.

+++ include/sampler.h
@@ -38,7 +38,7 @@ struct pmuevent {
 };
 
 struct logheader {
-  u32 ncpus;
+  u64 ncpus;
   struct {
     u64 offset;
     u64 size;
